\name{bread}
\alias{bread}
\docType{data}
\title{Breakfast preferences}
\description{The data set is described in Bro (1998). The raw data consist of ratings of 10 
breads on 11 different attributes carried out by 8 raters. 
Note that the bread samples are pairwise replications: Each of the 5 different breads, 
which have a different salt content, was presented twice for rating. 

}
\usage{data(bread)}
\format{A list of length 8 with elements of class \code{"dist"}. The attributes are 
bread odor, yeast odor, off-flavor, color, moisture, dough, salt taste, sweet taste, yeast taste, 
other taste, and total taste. 
}

\references{Bro, R. (1998). Multi-way Analysis in the Food Industry: Models, Algorithms, and Applications. Ph.D. thesis, University of Amsterdam (NL) \& Royal Veterinary and Agricultural University (DK).}
\examples{
bread
}
\keyword{datasets}
