\name{breakfast}
\alias{breakfast}
\docType{data}
\title{Breakfast preferences}
\description{42 individuals were asked to order 15 breakfast items due to their preference.
}
\usage{data(breakfast)}
\format{Data frame with students in the rows and breakfast items in the columns. 

V1: toast pop-up

V2: buttered toast

V3: English muffin and margarine

V4: jelly donut

V5: cinnamon toast

V6: blueberry muffin and margarine

V7: hard rolls and butter

V8: toast and marmalade

V9: buttered toast and jelly

V10: toast and margarine

V11: cinnamon bun

V12: Danish pastry

V13: glazed donut

V14: coffee cake

V15: corn muffin and butter

}

\references{Green, P. E. \& Rao, V. (1972). Applied multidimensional scaling. Hinsdale, IL: Dryden.
}
\examples{
data(breakfast)

}
\keyword{datasets}
