\name{CanadaNews}
\alias{CanadaNews}
\docType{data}
\title{Canadian Newspapers}
\description{We took Canadian newspapers that appeared in the time period between June and September 2009 and searched for articles that contained the word "aboriginal". A total of 92 articles was found. In these articles, we determined the frequencies of other meaningful words (e.g., tribal, moose, arctic, and health). The data are organized as word co-occurrence matrix.
}
\usage{data(CanadaNews)}
\format{Matrix with word co-occurrence counts.
}


\references{Borg, I., Groenen, P. J. F., & Mair, P. (2017). Applied Multidimensional Scaling and Unfolding. New York: Springer. 
}

\examples{
str(CanadaNews)
}
\keyword{datasets}
