\name{crimes}
\alias{crimes}
\docType{data}
\title{Crime Correlations}
\description{Correlations of crime rates in 50 US states.
}
\usage{data(crimes)}
\format{Crime correlation matrix.
}


\references{Borg, I., Groenen, P. J. F., & Mair, P. (2017). Applied Multidimensional Scaling and Unfolding. New York: Springer. 
}
\examples{
crimes
}
\keyword{datasets}
