\name{csrranking}
\alias{csrranking}
\docType{data}
\title{CSR activities}
\description{This dataset collects rankings of 100 individual on 5 topics that reflect social responsibilties on corporations. 
}
\usage{data(csrranking)}
\format{A data frame where each individual ranked prevention of environmental pollution (Environment), waste prevention (Waste Prevention),  selling organic products (Organic Products), participating on charity programs (Charity), and fair treatment of employees (Employee) according to its own preferences. A value of 1 corresponds to highest importance, 5 to lowest importance. 

}

\examples{
data(csrranking)

}
\keyword{datasets}
