\name{ekman}
\alias{ekman}
\docType{data}
\title{Ekman data set}
\description{Ekman dissimilarities
}
\usage{data(ekman)}
\format{Object of class \code{dist}
}
\details{Similarities of colors with wavelengths from 434 to 674 nm.}

\references{Ekman, G. (1954). Dimensions of color vision. Journal of Psychology, 38, 467-474.
}
\examples{
data(ekman)
## maybe str(ekman) ; plot(ekman) ...
}
\keyword{datasets}
