\name{Guerry}
\alias{Guerry}
\docType{data}
\title{Map Dataset France 1830}
\description{Airline distances (in km) between French Departments in 1830.
}
\usage{data(Guerry)}
\format{Symmetric matrix with distances. 
}

\references{Friendly, M. (2007). A. M. Guerry's Moral Statistics of France: Challenges for Multivariate Spatial Analysis, 
Statistical Science, 2007, 22(3), 368-399. 
}
\examples{
data(Guerry)
}
\keyword{datasets}
