\name{LawLer}
\alias{Lawler}
\docType{data}
\title{Management Performance Data}
\description{
Performance of managers: 3 criteria ("traits") and 3 methods. Traits:
T1 = Quality of output, T2 = Ability to generate output, T3 = Demonstrated effort to perform.
Methods: M1 = Rating by superior, M2 = Peer rating, M3 = Self-rating. 

}
\usage{data(Lawler)}
\format{Symmetric matrix (trait-method combinations) with inter-correlations. 
}

\references{Lawler, E. E. (1967).  Management performance as seen from above, below, and within.  In 
Evaluation of executive performance.  Princeton, New Jersey.  Educational Testing Service. 
}
\examples{
Lawler
}
\keyword{datasets}
