\name{partypref}
\alias{partypref}
\docType{data}
\title{Party preferences}
\description{Artificial dataset containing the judges in the rows and the parties in the columns.
}

\usage{data(partypref)}

\format{Matrix of party preferences.
}


\references{
Borg, I., Groenen, P. J. F., & Mair, P. (2010). Multidimensionale Skalierung. Muenchen: Hampp Verlag.

}
\examples{
partypref
}
\keyword{datasets}
