\name{perception}
\alias{perception}
\docType{data}
\title{Rectangle Perception Data}
\description{42 subjects are assigned to two groups of 21 persons. 120 stimulus pairs of rectangles are presented. For the first group (width-height; WH), the rectangles were constructed according to a design as given in \code{rect_constr}. For the second group (size-shape; SS) the rectangles were constructed according to a grid design, which is orthogonal in the dimensional system reflecting area (size), and width/height (shape). All subjects had to judge the similarity of the rectangles on a scale from 0 to 9. 
}

\usage{data(perception)}
\format{List of subject dissimilarities for WH (first element) and SS group (second element).
}

\references{Borg, I. & Leutner, D. (1983). Dimensional models for the perception of rectangles. Perception and Psychophysics, 34, 257-269.
}
\seealso{\code{\link{rectangles}}
}
\examples{
perception
rect_constr
}
\keyword{datasets}
