\name{RockHard}
\alias{RockHard}
\docType{data}
\title{RockHard Ratings}
\description{Data from RockHard Magazine: In this German Heavy Metal Magazine around 50 records are rated by the writers
on a scale from (0 ... worst to 10 ... best) each month. The dataset contains all ratings from 2013. 
}
\usage{data(RockHard)}
\format{Data frame with raters in the colums, bands/albums in the rows. 
}

\references{Mair, P., de Leeuw, J., & Wurzer, M. (2015). Multidimensional Unfolding. Wiley StatsRef: Statistics Reference Online. 
New York: Wiley.
}
\examples{
head(RockHard)
}
\keyword{datasets}
