\name{winedat}
\alias{winedat}
\docType{data}
\title{Wine tasting}
\description{This dataset collects dissimilarity matrices of 10 raters of 6 different wines.
}
\usage{data(winedat)}
\format{A list of dissimilarity matrices reflecting the rating of 10 judges on 6 different wines (Ziniel Chardonnay, Markowitsch Chardonnay, Krems Chardonnay, Castel Nova Chardonnay, Ritinitis Noble Retsina, RetsinaCriteria). The attributes color, smell, taste, fun, and overall impression were rated on a scale from 1 (very good) to 5. Based on these ratings the distances were computed.    
}

\examples{
data(winedat)

}
\keyword{datasets}
