\name{breakfast}
\alias{breakfast}
\docType{data}
\title{Breakfast preferences}
\description{42 individuals were asked to order 15 breakfast items due to their preference.
}
\usage{data(breakfast)}
\format{Data frame with students in the rows and breakfast items in the columns. 

toast: toast pop-up

butoast: buttered toast

engmuff: English muffin and margarine

jdonut: jelly donut

cintoast: cinnamon toast

bluemuff: blueberry muffin and margarine

hrolls: hard rolls and butter

toastmarm: toast and marmalade

butoastj: buttered toast and jelly

toastmarg: toast and margarine

cinbun: cinnamon bun

danpastry: Danish pastry

gdonut: glazed donut

cofcake: coffee cake

cornmuff: corn muffin and butter

}

\references{Green, P. E. \& Rao, V. (1972). Applied multidimensional scaling. Hinsdale, IL: Dryden.
}
\examples{
data(breakfast)

}
\keyword{datasets}
