\name{ekman}
\alias{ekman}
\docType{data}
\title{Ekman data set}
\description{Ekman dissimilarities
}
\usage{data(ekman)}
\format{Object of class \code{dist}
}
\details{Ekman presents similarities for 14 colors which are based on a rating by 31 subjects where each pair of colors 
was rated on a 5-point scale (0 = no similarity up to 4 = identical). 
After averaging, the similarities were divided by 4 such that they are within the 
unit interval. Similarities of colors with wavelengths from 434 to 674 nm.}

\references{Ekman, G. (1954). Dimensions of color vision. Journal of Psychology, 38, 467-474.
}
\examples{
ekman
}
\keyword{datasets}
