\name{indvalues}
\alias{indvalues}
\docType{data}
\title{Schwartz Value Survey}
\description{Responses from a sample in Britain were collected varying in value measures of the Schwartz value theory. The instrument used was the Schwartz Value Survey (SVS).
}

\usage{data(indvalues)}
\format{Data frame with 327 persons in the rows and psychological values in the columns. 
}

\details{
The data were centered (row-wise) and converted from preferences into dissimilarities.
}

\references{Borg, I., Bardi, A., & Schwartz, S. H. (2017). Does the value circle exist within persons or only across persons? Journal of Personality, 85(2), 151-162. 
}

\seealso{\code{\link{PVQ40}}}

\examples{
str(indvalues)
}
\keyword{datasets}
