\name{perception}
\alias{perception}
\docType{data}
\title{Rectangle Perception Data}
\description{42 subjects are assigned to two groups of 21 persons. 120 stimulus pairs of rectangles are presented. One group judges on a scale from 0 to 9 the perception in terms of width-height (WH), the other group in terms of size-shape (SS).  
}

\usage{data(perception)}
\format{List of subject dissimilarities for WH and SS group.
}

\references{Borg, I. & Leutner, D. (1983). Dimensional models for the perception of rectangles. Perception and Psychophysics, 34, 257-269.
}
\examples{
data(perception)

}
\keyword{datasets}
